+++ RELEASE-NOTES.md
@@ -1,5 +1,12 @@
 # SemRelease release notes
 
+## 2023-08-08 - v1.1.0
+
+- Add input `tag-prefix`.
+- Extend parsing to changelog file: `CHANGELOG.md`, `CHANGELOG.MD`, `CHANGELOG`, `CHANGE-LOG.md`, `CHANGE-LOG.MD`,
+`CHANGE-LOG`, `CHANGE_LOG.md`, `CHANGE_LOG.MD`, `CHANGE_LOG`, `changelog.md`, `changelog`, `change-log.md`,
+`change-log`, `change_log.md`, `change_log`.
+
 ## 2023-08-07 - v1.0.0
 
 - Publish releases using tags, following semantic versioning.